--- padertorch/contrib/neumann/dual_path_rnn/__init__.py
@@ -1,1 +1,0 @@
-
