+++ padertorch/models/__init__.py
@@ -1,4 +1,3 @@
 from . import bss
 from . import mask_estimator
 from . import wavenet
-from . import dual_path_rnn_source_separator
